util: Introduce a new LockAwareConfig interface
We keep KillbillConfig as a marker interface, and introduce LockAwareConfig to expose
common configs for global lock configuration.

In order to do this, we need to also introduce another base class MultiTenantLockAwareConfigBase to share
the implementation of such global lock configuration methods.
