+++ overdue/src/main/java/org/killbill/billing/overdue/config/MultiTenantOverdueConfig.java
@@ -27,10 +27,11 @@ import org.killbill.billing.util.config.definition.KillbillConfig;
 import org.killbill.billing.util.config.definition.OverdueConfig;
 import org.killbill.billing.util.config.tenant.CacheConfig;
 import org.killbill.billing.util.config.tenant.MultiTenantConfigBase;
+import org.killbill.billing.util.config.tenant.MultiTenantLockAwareConfigBase;
 import org.killbill.billing.util.glue.KillBillModule;
 import org.skife.config.TimeSpan;
 
-public class MultiTenantOverdueConfig extends MultiTenantConfigBase implements OverdueConfig {
+public class MultiTenantOverdueConfig extends MultiTenantLockAwareConfigBase implements OverdueConfig {
 
     private final OverdueConfig staticConfig;
